Task5

+++ app/src/main/java/com/nix/summer/final_project/core/entities/Resources.kt
@@ -5,5 +5,4 @@ data class Resources(
     val milk: Int = 540,
     val coffeeBeans: Int = 120,
     val cups: Int = 9,
-
-    val money: Int = 550)
+    val money: Float = 550f)
